added promotion to admin, edit categort and tags, final clean up

+++ Presentation/Controllers/SlotController.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
@@ -11,7 +9,6 @@ using BLL.Interfaces;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
 using Presentation.Models;
 
 namespace Presentation.Controllers
@@ -31,32 +28,28 @@ namespace Presentation.Controllers
             _mapper = mapper;
         }
 
-        // GET: api/Slot
         [HttpGet]
         [Route("page/{id:int}&{itemsOnPage}")]
         public async Task<Page> GetPage(int id, int itemsOnPage)
         {
-            if (id < 1)
+            if (id <= 0 || itemsOnPage <= 0)
             {
                 Response.StatusCode = 400;
+                await Response.WriteAsync("Page cannot be 0 or lower");
                 return null;
             }
-
-            var slotPage = await _slotRepresentationService.GetPage(id, itemsOnPage);
-
-            return slotPage;
+            return await _slotRepresentationService.GetPage(id, itemsOnPage);
         }
 
-        // GET: api/Slot/5
         [HttpGet("{id:int}")]
         public async Task<SlotFullDTO> GetSlot(int id)
         {
-            if (id < 1)
+            if (id <= 0)
             {
                 Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
                 return null;
             }
-
             try
             {
                 return await _slotRepresentationService.GetSlot(id);
@@ -74,37 +67,68 @@ namespace Presentation.Controllers
         }
 
         [HttpGet("{id}/price")]
-        public async Task<decimal> GetSlotPrice(int id)
+        public async Task<decimal?> GetSlotPrice(int id)
         {
-            return await _slotRepresentationService.GetSlotPrice(id);
+            if (id <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
+                return null;
+            }
+            try
+            {
+                var res = await _slotRepresentationService.GetSlotPrice(id);
+                Response.StatusCode = 200;
+
+                return res;
+            }
+            catch
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
         }
 
         [Authorize]
         [HttpGet("{id}/userBet")]
-        public async Task GetUserBet(int id)
+        public async Task<decimal?> GetUserBet(int id)
         {
-            var userId = Convert.ToInt32(User.Claims.First(x => x.Type == "Id").Value);
-            var res = await _slotRepresentationService.GetUserBet(id, userId);
-            if (res == -1)
+            if (id <= 0)
             {
-                return;
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
+                return null;
             }
+            var userId = Convert.ToInt32(User.Claims.First(x => x.Type == "Id").Value);
+            try
+            {
+                var res = await _slotRepresentationService.GetUserBet(id, userId);
+                Response.StatusCode = 200;
 
-            Response.StatusCode = 200;
-            await Response.WriteAsync(JsonConvert.SerializeObject(res,
-                new JsonSerializerSettings {Formatting = Formatting.Indented}));
+                return res;
+            }
+            catch
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
         }
 
-        // POST: api/Slot
         [Authorize]
         [HttpPost]
         public async Task Post([FromBody] SlotCreationModel newSlot)
         {
+            if (!ModelState.IsValid)
+            {
+                Response.StatusCode = 400;
+                await WriteErrors();
+                return;
+            }
             var id = Convert.ToInt32(User.Claims.First(x => x.Type == "Id").Value);
             try
             {
                 int slotId = await _slotManagementService.CreateSlot(id, _mapper.Map<SlotCreationDTO>(newSlot));
-                Response.StatusCode = 200;
+                Response.StatusCode = 201;
                 await Response.WriteAsync(slotId.ToString());
             }
             catch (Exception e)
@@ -114,31 +138,32 @@ namespace Presentation.Controllers
             }
         }
 
-        // POST: api/Slot
         [Authorize]
         [HttpPost]
         [Route("image/{id}")]
-        public async Task Post(IFormFile file, int id)
+        public async Task PostImage(IFormFile file, int id)
         {
+            if (id <= 0 || file == null)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower \n No file");
+                return;
+            }
             try
             {
                 if (!Directory.Exists(Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\"))
                 {
                     Directory.CreateDirectory(Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\");
                 }
-
                 using (FileStream filestream = System.IO.File.Create(
-                    Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\" +
-                    $"{Path.GetFileNameWithoutExtension(file.FileName)}_{id}{Path.GetExtension(file.FileName)}"))
+                    Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\" + $"_{id}{Path.GetExtension(file.FileName)}"))
                 {
                     file.CopyTo(filestream);
                     filestream.Flush();
                     await _slotManagementService.AddImageLink(id,
-                        $"/SlotImages/{Path.GetFileNameWithoutExtension(file.FileName)}_{id}{Path.GetExtension(file.FileName)}");
+                        $"/SlotImages/_{id}{Path.GetExtension(file.FileName)}");
                 }
-
-                Response.StatusCode = 200;
-                await Response.WriteAsync(id.ToString());
+                Response.StatusCode = 201;
             }
             catch (NotFoundException)
             {
@@ -154,14 +179,13 @@ namespace Presentation.Controllers
         [HttpPost("byCategory/{page}&{itemsOnPage}")]
         public async Task<Page> GetSlotsByCategory(int page, int itemsOnPage, [FromBody]int id)
         {
-            if (id < 0)
+            if (id <= 0 || page <= 0 || itemsOnPage <= 0)
             {
                 Response.StatusCode = 400;
+                await Response.WriteAsync("Page cannot be 0 or lower");
                 return null;
             }
-
             var slots = await _slotRepresentationService.GetByCategory(id, page, itemsOnPage);
-
             Response.StatusCode = 200;
             return slots;
         }
@@ -169,8 +193,13 @@ namespace Presentation.Controllers
         [HttpPost("byTags/{page}&{itemsOnPage}")]
         public async Task<Page> GetSlotsByTags(int page, int itemsOnPage, [FromBody]int[] ids)
         {
+            if (page <= 0 || itemsOnPage <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Page cannot be 0 or lower");
+                return null;
+            }
             var slots = await _slotRepresentationService.GetByTags(ids, page, itemsOnPage);
-
             Response.StatusCode = 200;
             return slots;
         }
@@ -178,8 +207,13 @@ namespace Presentation.Controllers
         [HttpPost("byName/{page}&{itemsOnPage}")]
         public async Task<Page> GetSlotsByName(int page, int itemsOnPage, [FromBody]string name)
         {
+            if (page <= 0 || itemsOnPage <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Page cannot be 0 or lower");
+                return null;
+            }
             var slots = await _slotRepresentationService.GetByName(name, page, itemsOnPage);
-
             Response.StatusCode = 200;
             return slots;
         }
@@ -188,6 +222,12 @@ namespace Presentation.Controllers
         [HttpPut("makeBet/{id}")]
         public async Task MakeBet(int id, [FromBody]decimal bet)
         {
+            if (id <= 0 || bet <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id or bet cannot be 0 or lower");
+                return;
+            }
             var userId = Convert.ToInt32(User.Claims.First(x => x.Type == "Id").Value);
             try
             {
@@ -204,6 +244,12 @@ namespace Presentation.Controllers
         [HttpDelete("undoBets/{id}")]
         public async Task UndoBet(int id)
         {
+            if (id <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
+                return;
+            }
             var userId = Convert.ToInt32(User.Claims.First(x => x.Type == "Id").Value);
             try
             {
@@ -216,14 +262,21 @@ namespace Presentation.Controllers
             }
         }
 
-        // PUT: api/Slot/5
         [Authorize]
         [HttpPut("{id}")]
-        public async Task Put(int id, [FromBody] SlotUpdateDTO slotInfo)
+        public async Task Put(int id, [FromBody] SlotUpdateModel slotInfo)
         {
+            if (id <= 0 || !ModelState.IsValid)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
+                await WriteErrors();
+                return;
+            }
             try
             {
-                await _slotManagementService.UpdateGeneralInfo(id, slotInfo);
+                var info = _mapper.Map<SlotUpdateDTO>(slotInfo);
+                await _slotManagementService.UpdateGeneralInfo(id, info);
                 Response.StatusCode = 204;
             }
             catch
@@ -236,29 +289,20 @@ namespace Presentation.Controllers
         [HttpPut("{id}/image")]
         public async Task UpdateImage(int id, IFormFile file)
         {
+            if (id <= 0 || file == null)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower \n No file");
+                return;
+            }
             try
             {
-                if (!Directory.Exists(Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\"))
-                {
-                    Directory.CreateDirectory(Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\");
-                }
-
-                using (FileStream filestream = System.IO.File.Create(
-                    Directory.GetCurrentDirectory() + "\\wwwroot\\SlotImages\\" +
-                    $"{Path.GetFileNameWithoutExtension(file.FileName)}_{id}{Path.GetExtension(file.FileName)}"))
-                {
-                    var slot = await _slotRepresentationService.GetSlot(id);
-                    string link = slot.ImageLink;
-                    link = Directory.GetCurrentDirectory() + "\\wwwroot" + link;
-                    System.IO.File.Delete(link.Replace("/", "\\"));
-                    file.CopyTo(filestream);
-                    filestream.Flush();
-                    await _slotManagementService.AddImageLink(id,
-                        $"/SlotImages/{Path.GetFileNameWithoutExtension(file.FileName)}_{id}{Path.GetExtension(file.FileName)}");
-                }
-
-                Response.StatusCode = 200;
-                await Response.WriteAsync(id.ToString());
+                var slot = await _slotRepresentationService.GetSlot(id);
+                string link = slot.ImageLink;
+                link = Directory.GetCurrentDirectory() + "\\wwwroot" + link;
+                System.IO.File.Delete(link.Replace("/", "\\"));
+                await PostImage(file, id);
+                Response.StatusCode = 204;
             }
             catch (NotFoundException)
             {
@@ -274,10 +318,45 @@ namespace Presentation.Controllers
         [HttpPut("{id}/status")]
         public async Task UpdateStatus(int id, [FromBody] string status)
         {
+            if (id > 0 && Enum.Parse(typeof(Status.SlotStatus), status) is Status.SlotStatus slotStatus)
+            {
+                try
+                {
+                    await _slotManagementService.UpdateStatus(id, slotStatus);
+                    Response.StatusCode = 204;
+                }
+                catch
+                {
+                    Response.StatusCode = 500;
+                }
+            }
+            else
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower \n No matching status found");
+            }
+        }
+
+        [Authorize]
+        [HttpDelete("{id}")]
+        public async Task Delete(int id)
+        {
+            if (id <= 0)
+            {
+                Response.StatusCode = 400;
+                await Response.WriteAsync("Slot id cannot be 0 or lower");
+                return;
+            }
+            var userId = Convert.ToInt32(User.FindFirst("Id").Value);
             try
             {
-                await _slotManagementService.UpdateStatus(id, (Status) Enum.Parse(typeof(Status), status));
-                Response.StatusCode = 204;
+                string imageLink = await _slotManagementService.DeleteSlot(id, userId);
+                if (imageLink != null)
+                {
+                    imageLink = Directory.GetCurrentDirectory() + "\\wwwroot" + imageLink;
+                    imageLink = imageLink.Replace("/", "\\");
+                    System.IO.File.Delete(imageLink);
+                }
             }
             catch
             {
@@ -285,16 +364,14 @@ namespace Presentation.Controllers
             }
         }
 
-        // DELETE: api/ApiWithActions/5
-        [Authorize]
-        [HttpDelete("{id}")]
-        public async Task Delete(int id)
+        private async Task WriteErrors()
         {
-            string imageLink = await _slotManagementService.DeleteSlot(id, Convert.ToInt32(User.FindFirst("Id").Value));
-            if (imageLink != null)
+            foreach (var modelStateValue in ModelState.Values)
             {
-                imageLink = Directory.GetCurrentDirectory() + "\\wwwroot" + imageLink;
-                System.IO.File.Delete(imageLink.Replace("/", "\\"));
+                foreach (var error in modelStateValue.Errors)
+                {
+                    await Response.WriteAsync(error.ErrorMessage);
+                }
             }
         }
     }
